dailyTasks-todo-app/pyproject.toml
[tool.poetry]
name = "dailytasks-todo-app"
version = "0.1.0"
description = ""
authors = ["Farhaj499 <farhajshauket@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlmodel = "^0.0.19"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pytest = "^8.2.2"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
